Move symlinking to kafka agent
To run in a shell do

$ conda activate /nsls2/conda/envs/2022-3.0-py39-tiled/
$ cd scripts/
$ python kafka_linker.py

and leave the window open
